feat: 마이페이지 닉네임 컴포넌트 분리
